requirements.txt
pyTelegramBotAPI==4.26.0
python-dotenv==1.1.0
openai==1.76.2